requirements-dev.txt
# extra
msgpack>=0.5.6
pyyaml>=3.13
tomli-w>=1.0
tomli>=1.1.0;python_version<'3.11'
orjson>=3.6.1

# tests
mypy>=0.812
flake8>=3.8.4
isort>=5.6.4
pytest>=6.2.1
pytest-mock>=3.5.1
pytest-cov>=2.10.1
coveralls>=3.0.0
flake8-isort>=4.0.0
black>=20.8b1
codespell>=2.2.2

# third party features
ciso8601>=2.1.3
pendulum>=2.1.2

# benchmark
pyperf>=2.6.1
termtables>=0.2.3
pytablewriter[html]>=0.58.0
cattrs==23.1.2
pydantic==2.1.1
dacite==1.7.0  # see https://github.com/konradhalas/dacite/issues/236#issuecomment-1613987368
marshmallow==3.20.1
dataclasses-json==0.5.14

# library stubs
types-backports
types-dataclasses
types-PyYAML
